backend/requirements.txt
Django
mysqlclient
djangorestframework
django-cors-headers
djangorestframework-simplejwt
PyJWT
